Update README.md
This repository contains the code for a PHP registration form that allows users to submit their information, which is then stored in a database. Users can also view the submitted data by clicking the "View Submitted Data" button and the data will be displayed on the tabular format with a unique registration number for every registration.